Absorb rasterizer into TileRenderer
